chore: move file handler utils to separate file
