+++ app/common/renderer/actions/Inspector.js
@@ -8,8 +8,8 @@ import i18n from '../i18next';
 import AppiumClient from '../lib/appium-client';
 import frameworks from '../lib/client-frameworks';
 import {getSetting, setSetting} from '../polyfills';
+import {readTextFromUploadedFiles} from '../utils/file-handling';
 import {getOptimalXPath, getSuggestedLocators} from '../utils/locator-generation';
-import {readTextFromUploadedFiles} from '../utils/other';
 import {
   findDOMNodeByPath,
   findJSONElementByPath,
