+++ app/common/renderer/utils/file-handling.js
@@ -1,0 +1,34 @@
+import {Promise} from 'bluebird';
+
+export function downloadFile(href, filename) {
+  let element = document.createElement('a');
+  element.setAttribute('href', href);
+  element.setAttribute('download', filename);
+  element.style.display = 'none';
+
+  document.body.appendChild(element);
+  element.click();
+
+  document.body.removeChild(element);
+}
+
+export async function readTextFromUploadedFiles(fileList) {
+  const fileReaderPromise = fileList.map((file) => {
+    const reader = new FileReader();
+    return new Promise((resolve) => {
+      reader.onload = (event) =>
+        resolve({
+          fileName: file.name,
+          content: event.target.result,
+        });
+      reader.onerror = (error) => {
+        resolve({
+          name: file.name,
+          error: error.message,
+        });
+      };
+      reader.readAsText(file);
+    });
+  });
+  return await Promise.all(fileReaderPromise);
+}
