+++ app/common/renderer/utils/other.js
@@ -1,4 +1,3 @@
-import {Promise} from 'bluebird';
 import _ from 'lodash';
 
 import {STANDARD_W3C_CAPS} from '../constants/session-builder';
@@ -46,37 +45,4 @@ export function parseCoordinates(element) {
   } else {
     return {};
   }
-}
-
-export function downloadFile(href, filename) {
-  let element = document.createElement('a');
-  element.setAttribute('href', href);
-  element.setAttribute('download', filename);
-  element.style.display = 'none';
-
-  document.body.appendChild(element);
-  element.click();
-
-  document.body.removeChild(element);
-}
-
-export async function readTextFromUploadedFiles(fileList) {
-  const fileReaderPromise = fileList.map((file) => {
-    const reader = new FileReader();
-    return new Promise((resolve) => {
-      reader.onload = (event) =>
-        resolve({
-          fileName: file.name,
-          content: event.target.result,
-        });
-      reader.onerror = (error) => {
-        resolve({
-          name: file.name,
-          error: error.message,
-        });
-      };
-      reader.readAsText(file);
-    });
-  });
-  return await Promise.all(fileReaderPromise);
 }
